// generated/src/app/api-models/pmb-subsc-bill-hist.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class PmbSubscBillHist {
  customerType: string;
  customerId: string;
  invoiceNo: number;
  subscriberId: string;
  recordType: string;
  planRiderCode: string;
  detailFromDate: Date;
  detailThruDate: Date;
  premiumAmt: number;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  premiumStep: string;
  familySize: number;
  spouseFlag: string;
}